requirements.txt
django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.9
gunicorn==20.1.0
redis==4.5.5
celery==5.3
boto3==1.26
python-dotenv==1.0.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==3.14.0
python-dateutil==2.8.2
django-filter==23.2
drf-yasg==1.21.5
mailersend==0.6.0